requirements.txt
fastapi[standard]
python-multipart
pymongo
python-dotenv
cloudinary
bcrypt
pyjwt
dnspython